Create BaseShell and use it for common 'default' and 'postcmd' methods
