+++ pepys_admin/admin_cli.py
@@ -7,6 +7,7 @@ from alembic import command
 from alembic.config import Config
 
 from paths import ROOT_DIRECTORY
+from pepys_admin.base_cli import BaseShell
 from pepys_admin.export_cli import ExportShell
 from pepys_admin.initialise_cli import InitialiseShell
 from pepys_admin.snapshot_cli import SnapshotShell
@@ -17,7 +18,7 @@ from pepys_import.utils.error_handling import handle_status_errors
 DIR_PATH = os.path.dirname(os.path.abspath(__file__))
 
 
-class AdminShell(cmd.Cmd):
+class AdminShell(BaseShell):
     """Main Shell of Pepys Admin."""
 
     intro = """--- Menu ---
@@ -125,20 +126,3 @@ class AdminShell(cmd.Cmd):
         """Exit the application"""
         print("Thank you for using Pepys Admin")
         sys.exit()
-
-    def default(self, line):
-        command_, arg, line = self.parseline(line)
-        # Python accepts letters, digits, and "_" character as a command.
-        # Therefore, "." is interpreted as an argument.
-        if arg == "." and line == ".":
-            self.aliases["."]()
-        elif command_ in self.aliases:
-            self.aliases[command_]()
-        else:
-            print(f"*** Unknown syntax: {line}")
-
-    def postcmd(self, stop, line):
-        if line != ".":
-            print("-" * 61)
-            print(self.intro)
-        return cmd.Cmd.postcmd(self, stop, line)
